[ADA] Port ADA to use Eigen + update to new sensors interface

See merge request scs/hermes/r2a-obsw!22

+++ src/boards/DeathStack/ADA/DeploymentUtils/elevation_map.cpp
@@ -6,7 +6,7 @@
 namespace elevationmap
 {
 
-int getElevation(double lat, double lon)
+int getElevation(float lat, float lon)
 {
     if (lat < SOUTH || lat >= NORTH || lon < WEST || lon >= EAST)
     {
