+++ src/boards/DeathStack/ADA/DeploymentUtils/elevation_map.h
@@ -30,6 +30,6 @@
 namespace elevationmap
 {
 
-int getElevation(double lat, double lon);
+int getElevation(float lat, float lon);
 
 } // namespace elevationmap
